First cut at support for parsing OpenSSH-compatible config files
This commit is a first cut at supporting OpenSSH-compatible config
files. Not all of the options are supported, and some options may never
be supported, but over time I'll continue to add support for the options
that have an equivalent AsyncSSH counterpart. So far, the following
client config options are supported:

  * Host conditional
  * Match conditional (all, host, originalhost, user, and localuser)
  * Include

  * AddressFamily
  * BindAddress
  * CASignatureAlgorithms
  * CertificateFile
  * Ciphers
  * Compression
  * ConnectTimeout
  * EnableSSHKeySign
  * ForwardAgent
  * ForwardX11Trusted
  * HostKeyAlgorithms
  * Hostname
  * IdentityAgent
  * IdentityFile
  * KexAlgorithms
  * MACs
  * Port
  * RekeyLimit
  * RemoteCommand
  * SendEnv
  * ServerAliveCountMax
  * ServerAliveInterval
  * SetEnv
  * User
  * UserKnownHostsFile

The following server config options are supported:

  * Match conditional (but only "all" so far)
  * Include

  * AddressFamily
  * BindAddress
  * Ciphers
  * ClientAliveCountMax
  * ClientAliveInterval
  * Compression
  * HostCertificate
  * HostKey
  * KexAlgorithms
  * MACs
  * Port
  * RekeyLimit

By default, client connections will read ~/.ssh/config unless you pass
in a different set of paths in the "config" argument. To completely
disable this new functionality, "config" can be set to None.

In any case where an option is set directly or via the existing
connection options classes, those values will take precedence over
anything file in a config file. One exception to this an initial
value of the host name to connect to can be set to a new value via
config, since host is always a required argument.

The existing keepalive_interval, login_timeout, rekey_bytes, and
rekey_seconds options now support the OpenSSH-style syntax such as
'1m30s' if you pass in string values for these options.

This commit also includes a few tests missed in a previous commit
related to passing in a separate list of certificates, and a fix in the
cleanup code of one of the SSH agent tests.

One bit of functionality not yet available is support for token
expansion. That will come in a future commit.