Add support for token percent expansion in new config module
This commit adds support for percent expansion of the following tokens
in a client config file:

    %%      literal %
    %C      SHA-1 hex digest of %l%h%p%r
    %d      Local user's home directory
    %h      Remote hostname
    %i      Local UID (UNIX only)
    %L      Local hostname (without the domain)
    %l      Local hostname (including the domain)
    %n      Original remote hostname
    %p      Remote port
    %r      Remote username
    %u      Local username

The "Hostname" config option supports %% and %h, while
"CertificateFile", "IdentityAgent", "IdentityFile", and
"RemoteCommand" support all of the above tokens.

Note that the %i expansion is not available on Windows. It only works
on systems which support os.getuid().

This commit also corrects 'Host' block matching on the client side to
always match against the original host passed in by the client, better
matching existing OpenSSH behavior. To match on the rewritten host, a
"Match host" block should be used.

This commit also prevents the User and Port config options from
being applied when the caller passes in an explicit username or port,
again better matching OpenSSH's behavior.