Add support for preferred client authentication methods
This commits adds support for an AsyncSSH client to choose a list of
authentication methods to use, in order of preference. When set,
the client will choose the first available method on the preferred list
which is supported by the server, and any methods not in the client's
preferred list will not be tried even if the server supports them.

This option is available as the "preferred_auth" argument in
SSHClientConnectionOptions or as the "PreferredAuthentications"
config file option.