Add support for ProxyJump config option
This commit adds support for the ProxyJump configuration option, as well
as allowing the "tunnel" argument to the various top-level connect and
listen functions to take a string identifying the host to tunnel
through, rather than requiring an already open SSHClientConnection. You
don't have as much control over options used to open the tunnel
conneciton when you pass in a string, but if all the options you need
are either specified in the config file or set by default, this is now
an option.

This commit also cleans up some unnecessary arguments in the forwarding
unit tests.