Continue to expand config file support
This commit adds support for the following options in both the client
and server config files:

    ChallengeResponseAuthentication
    HostbasedAuthentication
    KbdInteractiveAuthentication
    PasswordAuthentication
    PubkeyAuthentication

In addition to the above config file options, corresponding connection
options of "host_based_auth", "public_key_auth", "kbdint_auth", and
"password_auth" can be used to select which forms of auth are available.

The "ChallengeResponseAuthentication" option can be used as an
alternative to "KbdInteractiveAuthentication" to enable kbdint_auth,
but "ChallengeResponseAuthentication" will only be looked for if
"KbdInteractiveAuthentication" is not set.

All of these are enabled by default, but only if the client or server
provides the necessary input or callbacks to successfully complete that
form of authentication. However, these gestures can be used to disable
one or more of these forms of auth, even when information needed to
perform that auth is available.

This is useful to prevent the use of things like default public keys,
or to choose whether you want a password sent by password authentication
or keyboard-interactive authentication.

These options can also be used within Host/Match blocks to conditionally
enable/disable which forms of auth are supported.