Add support for RequestTTY and PermitTTY config options
This commit adds support for the RequestTTY client config option and the
corresponding PermitTTY server config option.

Previously, AsyncSSH teid the request of a pseudo-terminal to whether
the client specified a tereminal type or not. That behavior is still the
default, but it can be modified using a new "request_pty" connection
option, or the corresponding "RequestTTY" config option. All four of
the allowed OpenSSH values for this (yes, no, force, and auto) are
supported here, but "yes" is the default to match previous AsyncSSH
behavior, and the determination of whether stdin is a terminal is
replaced by whether a terminal type is set.

As part of this change, the existing server connection option
"allow_pty" is now associated with the server config option "PermitTTY".