Add server config file support for AuthorizedKeysFile
This commit adds support for the AuthorizedKeysFile server config
option, and extends the existing authorized_client_keys connection
option to support either a single file or a list of files.

This commit also adds a missing entry for the CASignatureAlgorithms
server config option. This was previously added to the client config,
but it really should have been supported for both client and server.