Add lists of the sent and received algorithms when matching fails
This commit enhances the "No matching algorithm found" exception which
can be raised during the key exchange to report the exact list of
algorithms which were sent and received to make it easier for the
caller to choose an algorithm that might fix the problem. It works
for multiple algorithm types, including key exchange, encryption,
MACs, compression, signature, and server host key algorithms. Thanks go
to Jeremy Schulman for suggesting this!