Add match and token substitution support for server config files
This commit adds support for "Match" conditional blocks in server config
files. Supported match types are LocalAddress, LocalPort, Host, User, and
Address, those Host is limited right now to only IP addresses
(consistent with OpenSSH having "UseDNS" set to "no", which is the
defualt).

OpenSSH also supports matching on Group, but since username in the
AsyncSSH server don't necessarily match local accounts on the host it
is running on, group lookups don't really make sense here.

OpenSSH also support RDomain (routing domain), but that is not supported
in AsyncSSH.

This commit also adds support for the '%u' substitution for user name in
the "AuthorizedKeyFile" config option, and fixes a bug in the handling
of 'Match all'support when such a block had a non-matching condition block
before it, and includes some other cleanup in config file support.