Add support for matching on reverse DNS lookup of client address
This commit adds support for the server config option "UseDNS", mapping
it to a new server connection option "rdns_lookup".  When set to True,
This option causes the SSH server code to do a reverse DNS lookup of
the client address, making that hostname available for matching using
the "from=" authorized keys option and the "Match Host" conditional
blocks in server config files. When "rdns_lookup" is set to False
(the default), these options can only match against IP addresses.