Fix issue with X.509 certificate-based authentication
This commit addresses an interoperability problem between AsyncSSH and
PKIXSSH, where PKIXSSH will reject a signature if the algorithm listed
in the public key being used for authentication doesn't match the
signature algorithm. With regular SSH keys OpenSSH certificates,
the algorithms may not always match. For instance, a key of type
"ssh-rsa" can support alternate signature algorithms such as
"rsa-sha2-256" and "rsa-sha2-512"."rsa-sha2-256" and "rsa-sha2-512".
Similarly, an "ssh-rsa" key might be used in an OpenSSH certificate,
but later the signature used might be "rsa-sha2-256" or "rsa-sha2-512".

In the case of X.509 certificates, signature algorithms of both
"x509v3-ssh-rsa" and "x509v3-rsa2048-sha256" are supported (with the
actual signature algorithm on the wire showing up as "ssh-rsa" or
"rsa2048-sha256"), but PKIXSSH rejects authentication if the
certificate's algorithm and the signature algorithm used to sign
the authentication request don't match.

With this change, the X.509 certificate data sent during authentication
is automatically adjusted so that the certificate's algorithm matches
the signature algorithm used to do the signing of the auth request.