Fix a couple of lurking bugs in unit test framework
I'm not sure how this wasn't noticed sooner, but this commit fixes a
couple of lurking bugs in setting up the SSH server used for doing
unit tests. One issue related to not properly setting up the certificate
in known_hosts, and another has to do with making sure Windows uses
the unit test framework's version of the known_hosts file.