Fix handling of alternate signature algorithms for PKCS#11 keys
This commit fixes a problem when AsyncSSH is configured to use an
alternate signature algorithms (such as rsa-sha2-256 vs. ssh-rsa).
A late change broke the non-certificate code path here, and the
unit tests only tried the primary signature algorithm.

This commit also extends the unit tests to test all available
signature algorithms for each of the supported key types, to make
sure future breakage of this sort is properly caught.