Update python-pkcs11 minimum version to 0.7.0
This commit updates the optional python-pkcs11 depedency in setup.py to
have a minimum version of 0.7.0. That version includes a fix which will
prevent that package from blocking other threads while waiting on calls
into the PKCS#11 provider. This allows the asyncio event loop to
continue running while the PKCS#11 call runs in a thread pool executor.