Minor updates to PKCS#11 support
This commit changes the implicit call to load_pkcs11_keys when using
the pkcs11_provider and pkcs11_pin connection options to retrieve the
keys using a thread pool executor, so that it doesn't block other
activity happening in the asyncio event loop.

This commit also takes advantage of existing functions in the pkcs11
module to filter on tokens and keys, rather than implementing that
in the AsyncSSH wrapper. This should be more efficient, particularly
with security tokens that contain a large number of keys.

Finally, this adds some notes in the documentation about when to
consider calling load_pkcs11_keys() explicitly vs. using the
pkcs11_provider and pkcs11_pin connection options and a suggestion
to consider using run_in_executor() when calling load_pkcs11_keys()
explicitly in an application where there is already an active event
loop.