Set exit status on SSH channel when acting as an SCP server
This commit sets the SSH exit status to 0 when cleanly closing inbound
SCP sessions when AsyncSSH is acting as an SCP server, to make sure
command-line SCP clients report a successful exit back to the shell.
Thanks go to girtsf for finding and reporting this, and identifying
the source of the problem!