Close local SSH listeners when SSH connection closes
This commit tracks listeners created with forward_local_port(),
forward_local_path(), or forward_socks() and automtically closes
them when the associated SSH connection closes, causing calls to
wait_closed() on the listener to return. Previously server-side
listeners were automatically closed and cleaned up on connection
close, but local listeners had to be clsoed explicitly by the
caller. Thanks go to rmawatson for reporting this issue and
suggesting a fix.

This commit also fixes an issue related to closing an automatically
opened tunnel connection when the tunneled connect inside it fails,
and fixes a cleanup-related race condition when a connection closes
unexpectedly while authentication is in progress.