Fix a race condition when receiving EOF immediately after channel open
This commit addresses a potential race condition between finishing the
open of a new channel and processing an incoming EOF message on that
channel. In some cases, the EOF was being processed before the channel's
session object was created, leading to an uncaught exception. With this
fix, EOF and close processing on a channel is deferred until the channel
startup completes, similar to how the code previously deferred processing
of received data. My thanks go to Alex Shafer for reporting this issue,
suggesting a possible fix, and helping me to track down the root cause!