Fix SFTP client error_handler to apply on initial stat failures
This commit enables the SFTP get/put/copy error handler to be invoked
when passing in a source path which doesn't exist or otherwise triggers
an error on the the initial stat()/lstat() call. Previously, such errors
were passed to the error handler in some cases (such as when recursion
was active), but not when the error occurred on an initial source path
provided by the caller. Thanks go to geraldnj for reporting this issue!