Call SFTP/SCP progress handler on empty files
This commit makes sure that the progress handler in SFTP and SCP will be
called even when empty files are transferred. Previously, this handler
was invoked after each block of data transferred, but that meant it
wasn't being called at all on empty files. Now, the progress handler is
called once on each empty file, reporting that 0 bytes were transferred.
Thanks again to geraldnj for noticing and reporting this.