Improve SCP support for pathlib path objects
This commit fixes an issue where a pathlib path object was not properly
handled as a source path, if it was passed in as a single path object
rather than a list of paths. It also fixes some issues with glob
matching not working right with pathlib objects.

This commit also improves unit testing of pathlib path objects for SCP
and refactors some of the unit test code to reduce duplication.