Avoid runtime error related to modifying dictionaries during iteration
This commit adds code in a few places to copy dictionary values before
iterating over them, in cases where the dictionary might be modified
during the call. This was already being done in many places, but was
missing in others. Thanks go to Thomas Léveil for reporting one of
these places and suggesting a fix.