Fix agent "confirm" unit test to work with OpenSSH 8.4p1 ssh-agent
This commit fixes the ssh-agent "confirm" unit test to work with
OpenSSH 8.4p1, which now requires that the DISPLAY variable not only
be present in the environment but that it not be an empty string before
it will run ssh-askpass. Previously, it checked only for the presence
of DISPLAY in the environment and not what it was set to.