Limit maximum user name length
This commit puts a limit on the maximum allowed length of a username, to
prevent a possible denial-of-service attack on AsyncSSH when it acts as
a server. The SASL prep code can take a significant amount of time to
run on long usernames, freezing up all other activity on the server
while this is being run.