Update X.509 cert generation to include subject & authority key IDs
For compliance with RFC 5280, this commit adds the SubjectKeyIdentifier
extension to all X.509 certificates generated by AsyncSSH and adds the
AuthorityKeyIdentifier extension to all non-self-signed certificates.

Subject key identifiers are derived from the public key in the certificate
being generated and authority key identifiers are derived from the public
key of the issuer.