Improve performance of line editor on long lines
This commit reimplements portions of the AsyncSSH line editor to make it
more efficient when processing long lines.

Unfortunately, even with these improvments, a very long input line can
still lead to the line editor preventing other connection handlers from
running for a significant amount of time. So, I'm planning to follow-up
this change with the ability to configure a maximum line length.

This doesn't really become an issue until lines are something like 100+
KB in size, so I don't expect most users to run into this. However,
having a configurable limit should help to prevent denial-of-service
attacks.