Add line editor input line length limit
This commit sets a 1024 byte length limit on input lines when the
AsyncSSH line editor feature is enabled, as well as a new server
connection option "max_line_length" which can be used to change
this value. This limit helps to prevent against denial of service
attacks when a client enters a very long input line into the line
editor.