Further improve line editor performance
This commit groups printable ASCII input in the line editor that has
no custom input handler and inserts an entire block of such input in
a single pass through the loop, rather than processing this input
one character at a time. It's currently limited to only matching on
printable plain ASCII since that can be matched with a very efficient
regex pattern, but it probably could be extended to support other
printable ranges if there was enough call for it.