Add support for "rmtree" method in SFTPClient
This commit adds support for an rmtree() method in the AsyncSSH SFTP
client which mirrors the rmtree() function in shutil, deleting all the
files in a directory tree.

A new scandir() method has also been added, return an async iterator
returning SFTPName entries for the contents of a directory. This allows
large directories to be scanned without having to load the complete
list of entries into memory via readdir() or listdir(). Instead, the
entries in a directory can be processed incrementally, periodically
going back to the SFTP server to get another batch of entries.

To speed things up, removals of different parts of the tree are
processed in parallel, and they begin even before the entire
contents of a directory are read using the new scandir() iterator.