Fix a couple of potential issues in SFTP rmtree()
This commit changes the STP rmtree() method to limit the number of
concurrent directory listing requests to avoid potentially hitting a
limit on the number of open file handles that SFTP servers allow in a
single session.

Also, exception handling has been improved to avoid potentially seeing
"Future exception was never retrieved" errors when multiple exceptions
are raised during its parallel execution. Only the first exception will
be reported back to the caller, but the others will be retrieved and
discarded. If a caller wants to see all the exceptions, they can do so
by specifying a callback in the "onerror" argument.