Add 'ignore_failure' in x11_forwarding option
This commit extends the x11_forwarding option in create_session() to
support a new value of 'ignore_failure' in addition to the existing
True and False values. Both True and 'ignore_failure' will request
X11 forwarding, but 'ignore_failure' ignores failures if they occur,
proceeding to open the new session as if forwarding was never
requested. This differs from True, which raises a ConnectionOpenError
on X11 forwarding failures.

This commit also changes the ForwardX11Trusted config file option to
behave as if x11_forwarding was set to 'ignore_failure' rather than
True, more closely matching OpenSSH's behavior for this option when
X11 forwarding can not be set up.