Make client channel options available as connection options
This commit makes it possible to set default values for all of
the session level options when opening a connection. Previously,
some of these values had fixed defaults and others had a default
which could be set via the recently added OpenSSH-compatible
config file, but it wasn't possible to explicitly set a default
value as part of the client connection options.

With this change, all session level options are now also connection
level options, so they can be set as arguments in a connect() call
or as values in SSHClientConnectionOptions.

Setting any of these values in create_session() takes precedence over
any value set in connect(), and any value set in connect() takes
precedence over any value set via the config file.

This change applies to the following options:

    command, subsystem, env, send_env, request_pty, term_type,
    term_size, term_modes, x11_forwarding, x11_display, x11_auth_path,
    x11_single_connection, encoding, errors, window, and max_pktsize