Extend support for replacing cert in SSHKeyPair
This commit refactors the code in SSHKeyPair to add support for
selecting an alternate certificate. Now, in addition to using this
on local key pairs, it can also be used on agent and PKCS11 keypairs.

Among other things, this means it is now possible to use an SSH agent
key with a certificate not loaded into the agent, including X.509
certificates. It also means that a PKCS11 key can be used with a
certificate not loaded onto the token, making it possible to use an
OpenSSH certificate.