Ignore unsupported algorithms in OpenSSH config file
This commit allows AsyncSSH to ignore unsupported algorithms present in
the OpenSSH config file, making it possible to configure algorithms
supported only by OpenSSH without breaking the ability for AsyncSSH to
share the config.

Strict checking is still done on any algorithms passed in as direct
arguments to the AsyncSSH API. This change only affects unsupported
algorithms listed in a config file.