Add method in SSH server classes to access terminal modes as a mapping
This commit adds a new get_terminal_modes() function to SSHServerChannel
which returns a mapping of all terminal modes set by the client when
starting up the session. It also adds new properties in SSHServerProcess
which can be used to access the terminal type, size, and modes, making
things more consistent with the way the client environment and other
process information was already available as properties.

With this commit, both the environment and the terminal modes are now
returned as MappingProxyType values, ensuring that the mappings cannot
be modified by the caller.

Thanks go to velavokr for submitting a PR requesting this change and
providing a proposed implementation.