Add support for ProxyCommand config option
This commit adds support for the ProxyCommand config file option, and a
corresponding proxy_command argument in the SSH connection options. This
is available in both the client and server options, as it can be used
with both connect() and connect_reverse(). It is also available as an
argument to get_server_host_key().

When set, a subprocess will be created running the requested command and
data will be piped to & from stdin/stdout on that process instead of
opening a TCP connection to the requested destination.

When using the config option, the usual OpenSSH percent substitutions
are available, such as %h for the destination host and %p for the
destination port.