Allow 'default' keyword for server_host_key_algs
This commit adds support for passing in 'default' as the value of
server_host_key_algs, as a way to request that the client send its
default set of supported host key algorithms to the server instead
of advertising only algorithms matching keys in its known_hosts.
This can allow known_host checking to be performed while avoiding
leaking information to the server about what types of host keys it
trusts.

This option should be used with caution, as it can lead to host key
mismatches if the client trusts only a subset of the host keys the
server might return.

Thanks go to Manfred Kaiser for suggesting this improvement!