Add new disable_trivial_auth argument to SSH client connection options
This commit adds a new client connection option "disable_trivial_auth".
When set, it prevents a client from connecting to a server which doesn't
perform some non-trivial form of authentication, such as requiring a
password or valid client public key. Thanks go to Manfred Kaiser for
suggesting this feature and providing an example implementation!