Remove host-based auth from the set of "non-trival" auth methods
This commit removes host-based auth from the set of authentication
methods that are allowed when disable_trivial_auth is set, since
that method doesn't involve any kind of challenge/prompt visible
to the user.

Also, this fixes a typo and reworks the wording somewhat in the
documentation for the server_host_key_algs argument.