Mask off high-order bits of SFTP permissions field
THe size of mode_t can vary between 16 and 32 bits on different OSes,
and even though anything outside of the bottom 16 bits is generally not
defined, some SFTP servers set these bits when encoding the permissions
as a UInt32, causing an overflow error in Python imlpementations when
mode_t is only 16 bits. This commit masks off those extra bits to avoid
the overflow error. Thanks go to GitHub user ccwufu for reporting this.