Allow encrypted keys to be ignored when no passphrase is set
This commit changes load_keypairs() to accept an option about whether or
not to ignore encrypted keys when no passphrase is specified. This
allows AsyncSSH to share a config file with OpenSSH with IdentityFile
references to encrypted private keys. If a passphrase is given to
AsyncSSH, these keys will be loaded, but otherwise the caller can choose
whether to ignore them or not with a new "ignore_encrypted" connection
option. This option is set by default when loading default keyapirs.

This commit also adds support for automatically ignoring some of the
default key files depending on whether or not the system has support for
the corresponding key types. If a key type is not supported, AsyncSSH
will skip attempting to load the associated default key file for that
key type.