Change SOCKS listener to return zero addresses instead of empty hostname
This commit changes the SOCKS5 listener to respond with an ATYP of either
IPv4 or IPv6 rather than an empty hostname which can apparently cause
problems for some SOCKS clients. Since the SSH client can't know the
assigned upsteam source IP and port, the response will have a zero value
for BND.ADDR and BND.PORT, but the response ATYP will be set to match
the request when that request asks to connect to an IPv4 or IPv6 address.
When the request is for a hostname, the response ATYP is set to IPv4
for maximum compatibility with older clients that might not be IPv6-aware.