Improve handling of forwarded connections which are immediately closed
This commit avoids an uncaught exception if a forwarded connection is
started but the socket associated with it is not filled into the
transport passed to connection_made(). This can apparently happen on
some systems, though so far I've been unable to reproduce that issue
myself.

This change guards against missing socket and peer name information
in the transport, expecting that the normal close process should
occur immediately after startup in this case.