Allow passphrase to be used to load keys used for SSH tunnels
This commit allows a passphrase specified to decrypt keys to be used
on connections through bastion hosts in addition to the main host. This
applies when using the ProxyJump config setting  as well as when using
the "tunnel" argument to specify a host and port to tunnel through.
Thanks go to Wout Megchelenbrink for reporting that this wasn't working
properly!