Add new connect_timeout connection option
This commit adds a new connect_timeout SSH connection option. It
applies to both client and server connections, whenever an outbound
TCP connection is being made. It also applies when opening a tunnel
which is then used to set up a client or server listener, and when
requesting SSH server host keys.

As part of this change, the existing ConnectTimeout config file option
now sets this new connect_timeout value instead of login_timeout,
which only applies after a TCP connection is established. The behavior
of this new option is more consistent with OpenSSH's ConnectTimeout
behavior.

A note has also been added to the documentation for login_timeout to
recommend the use of this new connect_timeout option when wanting
the timeout to include the time to establish the TCP connection.