Rework support for providing passphrase when opening tunnel
This commit reworks commit 16becd85, which accidentally broke passphrase
support on the primary connect function. This new implementation is
simpler and should properly allow the passphrase to be used for both
primary connections and a tunnel when that's not an already open
SSHClientConnection.

Note that other config options passed as arguments will not apply when
opening the tunnel, but any settings coming from the default config file
location will be applied when opening tunnels.

If you need more control over the options used when opening a tunnel
such as specifying different sets of keys or a different passphrase,
you should open the tunnel explicitly first and then pass the resulting
SSHClientConnection option as the tunnel argument.

This commit also includes a new set of unit tests which test passing
encrypted client keys with and without a passphrase (and with an invalid
passphrase) for both the primary connection and for tunnels.