Provide a better fix for the connect errors fixed in the last commit
This commit provides a better fix for the connect errors on Windows
that were corrected in the last commit. It turns out these were
related to a subtle difference in behavior between passing in None
vs. the empty string as the host when calling getaddrinfo() on
Windows.

This fix restores code that translates the empty string to None
internally when calling getaddrinfo(), but callers are expected to
always pass in a string and not pass None when using AsyncSSH APIs
to open a listener. This allows the type of the listen host argument
to be a simple str, rather than Optional[str].

This commit also updates places where None was being passed in the
unit tests to pass in an empty string and changes calls making direct
outbound connections to always pass an explicit address rather than
the empty string or None.

For backward compatibiity, callers not paying attention to type
annotations can pass in None and get the same behavior as before,
but an empty string is preferred.