Improve type annotations for async_context_maanger
This commit improves the type annotations for the async_context_manager
decorator.

This change uncovered an error in the type definitions in the classes
being used as async context maangers. This commit fixes this issue
and also removes the unused SCPFileProtocol class definition.

Note that mypy is not yet capable of properly enforcing the argument
types of the functions being passed into the async_context_manager
decorator. That would require support for ParamSpec, which mypy
doesn't have yet. Once it does, though, it will be a very small change
to take advantage of it.