Refactor SFTP glob support
This commit refactors SFTP glob support to reduce the number of
conversions of pathnames between bytes and strings when a str
or PurePath value is passed as a glob pattern to any of the copy
funcctions. With this change, a one-time conversion is done to
bytes and the expanded paths are kept as bytes for the remainder
of the operation.

Explicit calls to glob() will still convert the results back to
strings when a str or PurePath pattern is passed in, as will
calls to scandir, realpath, getcwd, and readlink. This change
only affects the behavior in the copy functions.

Along with this, the requirement that the destination be a directory
is now based on whether or not the number of source files after glob
expsnsion is greater than one. Previously, any use of mget/mput/mcopy
required the destination to be a directory, even if the glob pattern
matched only a single file.