Tighten up type hints on LocalFS
This commit narrows the type hints on the LocalFS class to reflect that
it only ever needs to deal with paths of type bytes in everything other
than the encode method. All paths are encoded as bytes before any of the
other functions are called.

This was already the case for SFTPFSProtocol and _SCPFSProtocol, which
are the protocol classes that LocalFS is designed to be compatible with.

This commit also renames SFTPFSProtocol to _SFTPFSProtocol since it is
not accessed from outside the module, and adds a missing "no cover"
pragma on a win32 conditional which defines _LocalPath.