Remove code which attempts to add extensions in FXP_INIT
Beginning with SFTPv4, extensions can no longer be sent by the client in
an FXP_INIT message. AsyncSSH never sent any such exceptions, but it did
parse (and then ignore) FXP_INIT extensions it received when operating as
a server. This commit removes code associated with sending such extensions
and limits parsing of such extensions on the server to only clients
requesting SFTPv3.