Finish support for end-of-list in FXP_NAME responses
This commit finishes adding support for the optional end-of-list value
in FXP_NAME responses, saving clients from having to send an additional
FXP_READDIR message to determine when they have reached the end of a
directory listing.

A prior commit parsed this value to avoid triggering an error, but never
acted upon it or sent it. With this new commit, full support for this
now exists when acting as either a client or a server. For efficiency,
the value is only added when it is True, and only in a response to an
FXP_READDIR request, even though other calls return FXP_NAME responses.