Revert name change in SFTPAttrs from nlink to link_count
While the "nlink" member of SFTPAttrs was previously undocumented and
only used internally when formatting the "longname" field in STPName
objects, there wasn't a lot of benefit in renaming it to "link_count"
(which better matches the SFTPv6 name). To avoid potentially
breaking application code which might have been setting nlink in
subclasses of SFTPServer, this commit returns to using the original
member name, now making that a publicly documented attribute.

This change has no effect to what is sent on the wire, where the
associated attribute bit is still FILEXFER_ATTR_LINK_COUNT (0x2000).