Reword client_keys documentation in SSHClientConnectionOptions
This commit updates the client_keys documentation to clarify the new
behavior introduced in AsyncsSH 2.5.0 when client_keys are configured
in conjunction with an ssh-agent or PKCS11 provider. In this case,
keys from all of these sources will be used. To get the old behavior
of using only client_keys, the agent_path and/or pkcs11_provider
parameters can be explicitly set to None.