Improve handling of task cancellation
This commit improves the handling of task cancellation in AsyncSSH,
making sure to cancel a future that it creates to wait for a connection
to be established if the application cancels a task before this wait
occurs. This avoids triggering an error of "Future exception was never
retrieved". Thanks go to Jonathan Tougas for reporting this issue and
providing test code to reliably reproduce it!