Update type annotation for SFTPClient open methods
This commit updates the type annotations on the return value of the
open() and open56() methods in SFTPClient to return an SFTPClientFile,
rather than an SFTPFileProtocol which declared only a subset of the
methods available in SFTPClientFile.

The protocol was being returned to avoid some mypy errors related to
the async_content_manager decorator, but this has now been fixed by
marking a type variable in that as "covariant".

Thanks go to Marat Sharafutdinov for reporting this - I had forgotten
that I hadn't resolved this before committing the initial set of type
annotations in an official release.