Update return types for LocalFS and SFTPServerFS open() methods
Mirroring the change in commit a4b2615, this commit updates the return
types in the open() call of the LocalFS and SFTPServerFS classes to
reflect the actual class they return rather than a protocol they
were created to follow.

This commit also fixes the async context manager methods in these
classes. While currently unused, the methods were defined to allow their
signature to match the open() method in SFTPClient. This change makes
sure that the async context manager funcitonality of these classes will
work correctly if it is ever used.