Fix a couple of config issues and add a debug log message
This commit fixes an issue with not properly handling path information
for nested includes in some cases, and not properly reporting when a
config include references a pattern with no matches.

This commit also adds a level 1 debug message whenever a config file is
loaded, to make it easier to see where config might be coming from.