Improve handling of config file loading when "options" argument is used
This commit attempts to improve the behavior of loading client config
files in the case where an "options" argument is used in the call to
connect(), listen_reverse(), or get_server_host_key().

Previously, if an options object was passed in on the connect() and no
config argument was specified when creating that object, the default SSH
config file would be loaded at the time the options object was created.
To avoid this, "config=None" had to be specified when creating the options.
Loading the config early prevented some conditionals from being evaluated
properly, since information to match those conditionals against wasn't
available yet. Also, leaving "config" unset when creating the options
object meant that "config=None" in the connect() call could not be honored
since the default config was already loaded before connect() was called.

With this change, the decision about whether to load the default config
or not is deferred until the connect() call when config is not specified
in either the options object creating or in the connect() call. This
delay also means that specifying "config=None" on the connect() call
now properly prevents the default config file from being loaded.